+++ packages/terra-application-docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 3.3.0 - (May 17, 2024)
+
 * Added
   * Added upgrade guide for terra-application v2.